refactor(Position): 패키징 수정
