+++ src/main/java/step3and4/car/Car.java
@@ -1,8 +1,7 @@
-package step3.car;
+package step3and4.car;
 
-import step3.number.Number;
-import step3.output.OutputView;
-import step3.position.Position;
+import step3and4.number.Number;
+import step3and4.output.OutputView;
 
 import java.util.Objects;
