+++ src/main/java/step3and4/cars/Cars.java
@@ -1,8 +1,8 @@
-package step3.cars;
+package step3and4.cars;
 
-import step3.car.Car;
-import step3.number.Number;
-import step3.output.OutputView;
+import step3and4.car.Car;
+import step3and4.number.Number;
+import step3and4.output.OutputView;
 
 import java.util.List;
 import java.util.Objects;
