+++ src/main/java/step3and4/number/RandomNumber.java
@@ -1,4 +1,4 @@
-package step3.number;
+package step3and4.number;
 
 import java.util.concurrent.ThreadLocalRandom;
