+++ src/test/java/step3and4/car/CarTest.java
@@ -1,11 +1,10 @@
-package step3.car;
+package step3and4.car;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.number.Number;
-import step3.output.OutputView;
-import step3.position.Position;
+import step3and4.number.Number;
+import step3and4.output.OutputView;
 
 class CarTest {
