+++ src/test/java/step3and4/cars/CarsTest.java
@@ -1,12 +1,12 @@
-package step3.cars;
+package step3and4.cars;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.car.Car;
-import step3.number.Number;
-import step3.number.RandomNumber;
-import step3.output.OutputView;
-import step3.position.Position;
+import step3and4.car.Car;
+import step3and4.car.Position;
+import step3and4.number.Number;
+import step3and4.number.RandomNumber;
+import step3and4.output.OutputView;
 
 import java.util.List;
